demo_test/pyproject.toml
[tool.poetry]
name = "demo-test"
version = "0.1.0"
description = ""
authors = ["Dmitry <tuma98@gmail.com>"]
readme = "README.md"
packages = [{include = "demo_test"}]

[tool.poetry.dependencies]
python = "^3.11"
appium-python-client = "^2.11.0"
appium-flutter-finder = "^0.4.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
